requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pillow>=9.0.0
albumentations>=1.3.0
tqdm>=4.64.0